requirements.txt
apparmor==3.0.4
apturl==0.5.2
asgiref==3.7.2
attrs==23.1.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.0
Django==4.2.7
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
fasteners==0.14.1
Flask==3.0.0
future==0.18.2
googlesearch-python==1.2.3
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
LibAppArmor==3.0.4
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==23.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.49
python-dotenv==0.21.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
ssh-import-id==5.11
systemd-python==234
texttable==1.6.7
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
websocket-client==0.59.0
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
